Add config option which controls SQS message deletion

+++ receiver/awss3receiver/s3sqsreader.go
@@ -64,6 +64,7 @@ type s3SQSNotificationReader struct {
 	s3Prefix            string
 	maxNumberOfMessages int32
 	waitTimeSeconds     int32
+	deleteMessages      bool
 }
 
 func newS3SQSReader(ctx context.Context, logger *zap.Logger, cfg *Config) (*s3SQSNotificationReader, error) {
@@ -92,6 +93,11 @@ func newS3SQSReader(ctx context.Context, logger *zap.Logger, cfg *Config) (*s3SQ
 		waitTime = int32(cfg.SQS.WaitTimeSeconds)
 	}
 
+	deleteMessages := true // Default to true
+	if cfg.SQS.DeleteMessages != nil {
+		deleteMessages = *cfg.SQS.DeleteMessages
+	}
+
 	return &s3SQSNotificationReader{
 		logger:              logger,
 		s3Client:            getObjectClient,
@@ -101,6 +107,7 @@ func newS3SQSReader(ctx context.Context, logger *zap.Logger, cfg *Config) (*s3SQ
 		s3Prefix:            cfg.S3Downloader.S3Prefix,
 		maxNumberOfMessages: maxMessages,
 		waitTimeSeconds:     waitTime,
+		deleteMessages:      deleteMessages,
 	}, nil
 }
 
@@ -213,12 +220,17 @@ func (r *s3SQSNotificationReader) readAll(ctx context.Context, _ string, callbac
 					}
 				}
 
-				_, err = r.sqsClient.DeleteMessage(ctx, &sqs.DeleteMessageInput{
-					QueueUrl:      aws.String(r.queueURL),
-					ReceiptHandle: message.ReceiptHandle,
-				})
-				if err != nil {
-					r.logger.Warn("Failed to delete message from SQS queue", zap.Error(err))
+				if r.deleteMessages {
+					r.logger.Debug("Deleting message from SQS queue",
+						zap.String("messageID", aws.ToString(message.MessageId)))
+
+					_, err = r.sqsClient.DeleteMessage(ctx, &sqs.DeleteMessageInput{
+						QueueUrl:      aws.String(r.queueURL),
+						ReceiptHandle: message.ReceiptHandle,
+					})
+					if err != nil {
+						r.logger.Warn("Failed to delete message from SQS queue", zap.Error(err))
+					}
 				}
 			}
 		}
